Split admin-firestore/index.ts's junk drawer into one-concept-per-file
index.ts (727 lines) mixed six unrelated responsibilities behind the
package's public barrel: the sandbox-only type surface (SandboxFirestore,
SnapshotObserver, Unsubscribe), the local-engine handle builder, the
rules-enforced getFirestore resolver, the rules-bypassing
getAdminFirestore resolver, and the full onSnapshot listener
implementation (four overloads, ref-shape discrimination, remote-arm
dispatch). Split into types.ts, local-handle.ts, get-firestore.ts,
get-admin-firestore.ts, and listeners.ts; index.ts is now a pure
re-export barrel with no function bodies, passing the barrel test in
docs/code-conventions.md section 4.

remote/remote-firestore.ts's SandboxFirestore import now points at
types.ts directly instead of round-tripping through the barrel.

Pure mechanical move: no behavior change. The package export
(pyric/sandbox/admin-firestore) and its re-exported surface are
unchanged.
